refactor: break out hooks managing all links page into it's own class
